[vgabios] Add 0x3802 IO port to shadow BDA data.
vgabios uses BDA to stash information that will later be used in BIOS
call (int10h), but BDA tends to be thrashed once the BIOS is done and
various sources tend to say it is not required to be preserved that long
anyway.

So add ioports 0x3800->0x383F to those informations.
See use in vgabios patch.